pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "comfyui-worker"
version = "0.1.0"
description = "LittleHorse worker sidecar for ComfyUI workflows."
requires-python = ">=3.11"
dependencies = [
  "httpx>=0.27.0",
  "littlehorse-client>=0.16.0",
  "pydantic>=2.6.0",
]

[dependency-groups]
dev = [
  "pytest-anyio>=0.0.0",
  "pytest-httpx>=0.30.0",
  "pytest>=8.0.0",
  "ruff>=0.14.13",
]

[tool.hatch.build.targets.wheel]
packages = ["comfyui_worker"]


[tool.pytest.ini_options]
